# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Aurélien",
        "description": "C très Bo Whoulla",
        "contact": {
            "email": "admin9@fromage.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "CRUD_Genre"
                ],
                "summary": "Get all genres",
                "description": "Returns all genres",
                "operationId": "getAllGenres",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "post": {
                "tags": [
                    "CRUD_Genre"
                ],
                "summary": "Create genre",
                "description": "create genre",
                "operationId": "createGenre",
                "requestBody": {
                    "description": "Pass genre information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Store a newly created resource in storage.",
                                        "type": "string",
                                        "example": "Hentai"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/genres/{id}": {
            "get": {
                "tags": [
                    "CRUD_Genre"
                ],
                "summary": "Get genre by ID",
                "description": "Returns genre by his id",
                "operationId": "getGenreByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "genre id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Id not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CRUD_Genre"
                ],
                "summary": "Delete genre by ID",
                "description": "Delete genre by his id",
                "operationId": "deleteGenreByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "genre id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "genre was deleted",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            },
            "patch": {
                "tags": [
                    "CRUD_Genre"
                ],
                "summary": "Update genre by ID",
                "description": "Update genre by his id",
                "operationId": "updateGenreByID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "genre id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Pass genre information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Update the specified resource in storage.",
                                        "type": "string",
                                        "example": "NUUUUUUUUUUL"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "Id not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/register": {
            "post": {
                "tags": [
                    "CRUD_User"
                ],
                "summary": "Register",
                "description": "create user",
                "operationId": "createUser",
                "requestBody": {
                    "description": "Pass user information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "c_password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Register api",
                                        "type": "string",
                                        "example": "THIS IS A NAME"
                                    },
                                    "email": {
                                        "description": "Register api",
                                        "type": "string",
                                        "format": "email",
                                        "example": "email@mmmmmmmmmmmmaiiiiiiiiil.mail"
                                    },
                                    "password": {
                                        "description": "Register api",
                                        "type": "string",
                                        "format": "password",
                                        "example": "000_yes_its_true"
                                    },
                                    "c_password": {
                                        "description": "Register api",
                                        "type": "string",
                                        "format": "password",
                                        "example": "000_yes_its_true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        },
        "/login": {
            "post": {
                "tags": [
                    "CRUD_User"
                ],
                "summary": "Login",
                "description": "login",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user information",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "Login api",
                                        "type": "string",
                                        "format": "email",
                                        "example": "email@mmmmmmmmmmmmaiiiiiiiiil.mail"
                                    },
                                    "password": {
                                        "description": "Login api",
                                        "type": "string",
                                        "format": "password",
                                        "example": "000_yes_its_true"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful created",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "scalar": "bearerAuth"
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
    "security": [
        []
    ]
}